fix: divide file for ibc test setting
